chore: move composeMiddlewares to middlewares folder
