+++ apps/core/middleware.ts
@@ -1,6 +1,6 @@
+import { composeMiddlewares } from './middlewares/compose-middlewares';
 import { withCustomUrls } from './middlewares/with-custom-urls';
 import { withMaintenanceMode } from './middlewares/with-maintenance-mode';
-import { composeMiddlewares } from './utils/composeMiddlewares';
 
 export const middleware = composeMiddlewares(withMaintenanceMode, withCustomUrls);
